pyproject.toml
[tool.poetry]
name = "opendesktopagent"
version = "0.1.0"
description = ""
authors = ["gr-b@github.com"]
license = "n"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.3"
Flask-Cors = "^3.0.10"
ipdb = "^0.13.13"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
gunicorn = "^21.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
